Split gambling games into 4 different commands
